requirements-pip.txt
numpy==1.26.*
scipy==1.12.*
joblib==1.3.*
scikit-learn==1.4.*
matplotlib==3.8.*
pillow==10.*
gpytorch==1.11
lightgbm>=4.3
